+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Changelog
+## [0.0.1] - 2020-05-22
+- Add in JOMI Combo stroke: ksc: ctrl + a + j
+- Add plenty of Window Macro In JavaOMI to use
+- Experiment sample structure in the package
+
 ## [0.0.1] - 2020-05-21
 - Log Start ;)